Reorganize remaining components into folders
