+++ src/components/MainDashboard/index.js
@@ -1,6 +1,7 @@
 export * from './FilterTopbar'
 export * from './FilteredComponent'
 export * from './NoEventsFound'
-export * from './EventList'
-export * from './OrganizationCard'
-export * from './OrganizationsList'
+export * from './EventView/EventList'
+export * from './EventView/Event'
+export * from './Orgview/OrganizationCard'
+export * from './Orgview/OrganizationsList'
