+++ src/components/Messages/index.js
@@ -1,0 +1,2 @@
+export * from './MessageThreads'
+export * from './Messages'
