+++ src/components/SiteParts/index.js
@@ -1,0 +1,3 @@
+export * from './FooterDiv'
+export * from './HeaderDiv'
+export * from './Navigation'
