+++ src/components/UtilComponents/index.js
@@ -1,0 +1,3 @@
+export * from './ProgressBar'
+export * from './ScrollToTop'
+export * from './UploadImage'
