+++ src/components/index.js
@@ -1,12 +1,10 @@
-export * from './Event';
-export * from './MainDashboard/EventList';
-export * from './Navigation';
-export * from '../styled/StyledUploadImage';
-export * from './FooterDiv';
-export * from './HeaderDiv';
-export * from './UploadImage';
-export * from './MainDashboard/FilteredComponent';
-export * from './MainDashboard/FilterTopbar';
-export * from './MainDashboard/NoEventsFound';
-export * from './ProgressBar';
-export * from '../styled/StyledLoader';
+export * from './CreateEvent'
+export * from './CreateOrg'
+export * from './EventComments'
+export * from './LandingPage'
+export * from './MainDashboard'
+export * from './Messages'
+export * from './OrgDashboard'
+export * from './SiteParts'
+export * from './UserProfile'
+export * from './UtilComponents'
